+++ tests/__main__.py
@@ -8,6 +8,7 @@ from tests.test_deserialization_from_sympy import *
 from tests.test_add_simplification import *
 from tests.test_mul_simplification import *
 from tests.test_as_decimal import *
+from tests.test_pi import *
 
 
 if __name__ == "__main__":